Create an initial version of the CLI
The NewProject command creates a new NerdPress project directory structure
with the necessary files to get started. This generator is more or less
complete with this commit, though there will be some updates as the
library evolves.

The Build command will process the source files and export the project in
the requested formats: all (default), HTML, PDF, EPUB, MOBI, or any
combination thereof.