Implement Section HTML output and export
The system will support a chained workflow of individual steps via the
Processor abstraction. Each Processor will act as a service object that
will work on a string of input and return a transformed output string.

In this initial implementation, the Processor workflow is empty so the
output equals the input.