Add Stylesheet abstraction
The Stylesheet represents a CSS or Sass file which is used to affect
the look and feel of the exported ebooks. The toolchain will move
all styles to the appropriate folder in the build directory (compiling
Sass stylesheets to CSS in the process) and use them when exporting
versions of the project based on the name of the format being exectuted.
(Ex: if you want to export a PDF version, the toolchain will expect you
to have a pdf.(css|sass|scss) file in the stylesheets source directory.
Common styles can be stored in non-format-specific stylesheet files and
imported into Sass-based stylesheets for reuse.